+++ packages/wiz-ui/src/components/base/header/header.stories.ts
@@ -4,7 +4,7 @@ import { SPACING_ACCESSORS } from "@wizleap-inc/wiz-ui-constants/styles/spacing"
 import { provide } from "vue";
 
 import { WizTextButton } from "@/components";
-import { globalKey, useGlobalProvider } from "@/providers";
+import { globalKey, useGlobalProvider } from "@/hooks/use-global-provider";
 
 import { WizHeader } from ".";
